+++ decoder/api_pokemon_scan_v3.go
@@ -204,7 +204,7 @@ func GrpcGetPokemonInArea3(retrieveParameters *pb.PokemonScanRequestV3) ([]*pb.P
 		if pokemon != nil {
 			if pokemon.ExpireTimestamp.ValueOrZero() > startUnix {
 				apiPokemon := pb.PokemonDetails{
-					Id:         pokemon.Id,
+					Id:         uint64(pokemon.Id),
 					PokestopId: pokemon.PokestopId.Ptr(),
 					SpawnId:    pokemon.SpawnId.Ptr(),
 					Lat:        pokemon.Lat,
